LayoutTests/media/modern-media-controls/media-controller/media-controller-tight-padding.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<style type="text/css" media="screen">
    
    audio, #shadow {
        position: absolute;
        left: 0;
        top: 0;
        width: 400px;
        height: 300px;
    }
    
</style>
<body>
<audio src="../../content/test.mp4" controls autoplay></audio>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>MediaController</code> sizing behavior with tight padding mode.");

const shadowRoot = document.querySelector("div#shadow").attachShadow({ mode: "open" });
const media = document.querySelector("audio");
const mediaController = createControls(shadowRoot, media, null);

shouldBecomeEqual("!!(mediaController.controls.layoutTraits & LayoutTraits.TightPadding)", "true", finishJSTest);

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
